+++ pakyow-core/spec/features/common/support/project/apps/test_1/backend/controllers/root.rb
@@ -1,0 +1,5 @@
+controller :root do
+  def self.context
+    :application
+  end
+end